backend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0

# Web Scraping
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==5.0.0

# AI / OpenAI
openai==1.6.1

# Database
supabase==2.3.0
python-dotenv==1.0.0

# Utilities
requests==2.31.0
pydantic==2.5.3
python-dateutil==2.8.2
